Cast size to double
An implicit conversion warning was being hidden by the use of -isystem
to include frameworks/native/include. Making changes to art trigered
the warning, which led to -Werror breaking build. Rather than disabling
-Werror, lets cast size to a double so we have correct precision.

+++ libs/ui/GraphicBufferAllocator.cpp
@@ -56,7 +56,7 @@ void GraphicBufferAllocator::dump(String8& result) const
         const alloc_rec_t& rec(list.valueAt(i));
         if (rec.size) {
             snprintf(buffer, SIZE, "%10p: %7.2f KiB | %4u (%4u) x %4u | %8X | 0x%08x | %s\n",
-                    list.keyAt(i), rec.size/1024.0f,
+                    list.keyAt(i), static_cast<double>(rec.size/1024.0f),
                     rec.width, rec.stride, rec.height, rec.format, rec.usage,
                     rec.requestorName.c_str());
         } else {
@@ -68,7 +68,7 @@ void GraphicBufferAllocator::dump(String8& result) const
         result.append(buffer);
         total += rec.size;
     }
-    snprintf(buffer, SIZE, "Total allocated (estimate): %.2f KB\n", total/1024.0f);
+    snprintf(buffer, SIZE, "Total allocated (estimate): %.2f KB\n", static_cast<double>(total/1024.0f));
     result.append(buffer);
     std::string deviceDump = mDevice->dump();
     result.append(deviceDump.c_str(), deviceDump.size());
